+++ www/main.js
@@ -1,0 +1,8 @@
+
+function do_nothing(){
+    console.log("do_nothing");
+}
+
+function set_copy(){
+    document.getElementById("footer").innerHTML = "<div>&copy; 2020 <a target='_blank' href='https://github.com/kr-g/mpymodcore'>modcore / homepage</a></div>";
+}